+++ packages/components/button/src/Button.stories.tsx
@@ -1,59 +1,76 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { fn } from '@storybook/test';
+import type { Meta, ReactRenderer, StoryObj } from '@storybook/react';
 import Button from './Button';
+import { ButtonProps } from './button.types';
 import { ThemeProvider } from '@douro-ui/react';
+import { PartialStoryFn } from 'storybook/internal/types';
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
-const meta = {
+const meta: Meta<ButtonProps> = {
   title: 'Example/Button',
-
-  component: () => (
-    <ThemeProvider>
-      <Button label="Button" />
-    </ThemeProvider>
-  ),
+  component: Button,
+  decorators: [
+    (Story: PartialStoryFn<ReactRenderer, ButtonProps>) => (
+      <ThemeProvider>
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
   parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
     layout: 'centered',
   },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ['autodocs'],
-  // More on argTypes: https://storybook.js.org/docs/api/argtypes
-  argTypes: {
-    backgroundColor: { control: 'color' },
+  args: {
+    size: 'lg',
+    disabled: false,
+    typeBtn: 'primary',
+    children: 'Button',
   },
-  // Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
-  args: { onClick: fn() },
-} satisfies Meta<typeof Button>;
+  argTypes: {
+    typeBtn: {
+      control: {
+        type: 'select',
+        options: ['primary', 'secondary', 'tertiary', 'error'],
+      },
+    },
+    size: {
+      control: { options: ['sm', 'md', 'lg', 'xl'] },
+    },
+    disabled: {
+      control: 'boolean',
+    },
+  },
+} satisfies Meta<ButtonProps>;
 
 export default meta;
-type Story = StoryObj<typeof meta>;
 
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
+type Story = StoryObj<ButtonProps>;
+
 export const Primary: Story = {
   args: {
-    primary: true,
-    label: 'Button',
-    backgroundColor: '#000000',
+    typeBtn: 'primary',
   },
 };
 
 export const Secondary: Story = {
   args: {
-    label: 'Button',
+    typeBtn: 'secondary',
   },
 };
 
-export const Large: Story = {
+export const Tertiary: Story = {
   args: {
-    size: 'large',
-    label: 'Button',
+    typeBtn: 'tertiary',
   },
 };
 
-export const Small: Story = {
+export const Error: Story = {
   args: {
-    size: 'small',
-    label: 'Button',
+    typeBtn: 'error',
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    typeBtn: 'primary',
+    disabled: true,
   },
 };