+++ packages/components/button/src/button.types.ts
@@ -1,22 +1,26 @@
-export interface ButtonProps {
-  /**
-   * Is this the principal call to action on the page?
-   */
-  primary?: boolean;
-  /**
-   * What background color to use
-   */
-  backgroundColor?: string;
-  /**
-   * How large should the button be?
-   */
-  size?: 'small' | 'medium' | 'large';
-  /**
-   * Button contents
-   */
-  label: string;
-  /**
-   * Optional click handler
-   */
+import { ButtonHTMLAttributes, ReactNode } from 'react';
+
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  typeBtn?: 'primary' | 'secondary' | 'tertiary' | 'error';
+  size?: 'sm' | 'md' | 'lg' | 'xl';
+  disabled?: boolean;
+  children?: ReactNode;
   onClick?: () => void;
+  styled?: ButtonStyledProps;
+}
+
+export interface ButtonStyledProps {
+  color?: string;
+  colorHover?: string;
+  colorActive?: string;
+  colorDisabled?: string;
+  backgroundColor?: string;
+  backgroundColorHover?: string;
+  backgroundColorActive?: string;
+  backgroundColorDisabled?: string;
+  borderColor?: string;
+  borderColorHover?: string;
+  borderColorActive?: string;
+  borderColorDisabled?: string;
+  borderRadius?: string;
 }