+++ packages/components/button/src/buttonTypes/index.ts
@@ -1,0 +1,4 @@
+export { PrimaryButton } from './PrimaryButton';
+export { SecondaryButton } from './SecondaryButton';
+export { TertiaryButton } from './TertiaryButton';
+export { ErrorButton } from './ErrorButton';