+++ packages/react/src/context/ThemeProvider.tsx
@@ -1,17 +1,16 @@
 import { ThemeProvider as ThemeProviderEmotion } from '@emotion/react';
-import React from 'react';
+import React, { ReactElement } from 'react';
 import type { Theme } from '../theme/theme.types';
 import { deepMerge } from '../utils/deepMerge';
 import { defaultTheme } from '../theme';
-import { EmotionJSX } from '@emotion/react/types/jsx-namespace';
 
 export const ThemeProvider = ({
   theme,
   children,
 }: {
-  theme?: Partial<Theme>;
+  theme?: Theme;
   children: React.ReactNode;
-}): EmotionJSX.Element => {
+}): ReactElement => {
   const mergedTheme = deepMerge({ ...defaultTheme }, theme);
   return (
     <ThemeProviderEmotion theme={mergedTheme}>{children}</ThemeProviderEmotion>