+++ packages/react/src/index.ts
@@ -1,3 +1,4 @@
 export * from './hooks';
 export * from './context';
 export * from './theme';
+export * from './utils';