+++ packages/react/src/theme/theme.types.ts
@@ -36,10 +36,10 @@ export type FontFamilyType<FontFamily extends string> = {
 
 export interface Theme {
   name: string;
-  fontSize: number;
+  fontSize: string;
   fontFamily: FontFamilyType<'display' | 'text'>;
   spaceUnit: Record<string, string>;
-  fontWeight: typeof FontWeight;
+  fontWeight: { [K in keyof typeof FontWeight]: FontWeight };
   colors: Colors;
   typography: Typography;
   breakpoints: BreakpointsType;