+++ packages/components/button/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.2.0...@douro-ui/button@0.2.1) (2024-07-26)
+
+### Bug Fixes
+
+- button without props ([3d39b88](https://github.com/Douro-ui/design-system/commit/3d39b88e503031ec760da980bd5197a7d5b24417))
+
+# [0.2.0](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.1.8...@douro-ui/button@0.2.0) (2024-07-25)
+
+### Features
+
+- button component ([c68aaee](https://github.com/Douro-ui/design-system/commit/c68aaeeeaff3e44bf905ba9335d828bbed11fe10))
+
+## [0.1.8](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.1.7...@douro-ui/button@0.1.8) (2024-07-24)
+
+**Note:** Version bump only for package @douro-ui/button
+
+## [0.1.7](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.1.6...@douro-ui/button@0.1.7) (2024-07-23)
+
+**Note:** Version bump only for package @douro-ui/button
+
+## [0.1.6](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.1.5...@douro-ui/button@0.1.6) (2024-07-18)
+
+**Note:** Version bump only for package @douro-ui/button
+
 ## [0.1.5](https://github.com/Douro-ui/design-system/compare/@douro-ui/button@0.1.4...@douro-ui/button@0.1.5) (2024-07-16)
 
 ### Bug Fixes