+++ packages/components/checkbox/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/Douro-ui/design-system/compare/@douro-ui/checkbox@0.1.6...@douro-ui/checkbox@0.1.7) (2024-07-24)
+
+**Note:** Version bump only for package @douro-ui/checkbox
+
+## [0.1.6](https://github.com/Douro-ui/design-system/compare/@douro-ui/checkbox@0.1.5...@douro-ui/checkbox@0.1.6) (2024-07-23)
+
+**Note:** Version bump only for package @douro-ui/checkbox
+
+## [0.1.5](https://github.com/Douro-ui/design-system/compare/@douro-ui/checkbox@0.1.4...@douro-ui/checkbox@0.1.5) (2024-07-18)
+
+**Note:** Version bump only for package @douro-ui/checkbox
+
 ## [0.1.4](https://github.com/Douro-ui/design-system/compare/@douro-ui/checkbox@0.1.3...@douro-ui/checkbox@0.1.4) (2024-07-16)
 
 ### Bug Fixes